Updated data model to a general game struct.
This allows for 1v1, 1v2, and 2v2 games by simply repeating the single players id in
both the offense and defense fields. This model also has the advantage of keeping
track of what position (offense or defense) a player played for any given game,
which could allow for statistics based on that.

Also added a tagline field to the player struct